+++ lib/src/visitor/evaluate.dart
@@ -5,7 +5,7 @@
 // DO NOT EDIT. This file was generated from async_evaluate.dart.
 // See tool/grind/synchronize.dart for details.
 //
-// Checksum: faf491d48ccd341abf6a301773bfce08af22b113
+// Checksum: a92289df39e6f2dbd248385a82d40456aadf8a6c
 //
 // ignore_for_file: unused_import
 
@@ -993,6 +993,7 @@ final class _EvaluateVisitor
   /// that they don't modify [root] or its dependencies.
   CssStylesheet _combineCss(Module<Callable> root, {bool clone = false}) {
     if (!root.upstream.any((module) => module.transitivelyContainsCss)) {
+      root.extensionStore.trimModernSelectors();
       var selectors = root.extensionStore.simpleSelectors;
       if (root.extensionStore
               .extensionsWhereTarget((target) => !selectors.contains(target))
@@ -1083,6 +1084,7 @@ final class _EvaluateVisitor
         module.extensionStore.addExtensions,
       );
       if (module.extensionStore.isEmpty) continue;
+      module.extensionStore.trimModernSelectors();
 
       for (var upstream in module.upstream) {
         if (upstream.url case var url?) {